get_daily_data/requirements.txt
pandas==0.23.3
google-cloud-bigquery==1.9.0
google-cloud-secret-manager==0.2.0
google-cloud-storage==1.13.2
numpy==1.18.1
pyarrow==0.12.0
requests==2.22.0
ta==0.5.11